requirements.txt
django==4.2.4
djangorestframework==3.14.0
requests==2.31.0
mysqlclient==2.2.0
elasticsearch==7.13.4
docker==6.1.3
scikit-learn==1.3.0